+++ database/migrations/2023_02_04_162450_create_maintenance_costs_table.php
@@ -15,8 +15,8 @@ return new class extends Migration
     {
         Schema::create('maintenance_costs', function (Blueprint $table) {
             $table->id();
-            $table->integer('cleaning');
-            $table->integer('accessories');
+            $table->decimal('cleaning', 10, 2);
+            $table->decimal('accessories', 10, 2);
             $table->unsignedBigInteger('automobile_cost_id');
             $table->foreign('automobile_cost_id')->references('id')->on('automobile_costs')->onDelete('cascade');
             $table->timestamps();
